Feature/instance name (#2)
* Add instance name variable

* Update readme to add instance name variable addition

* Add elastic ip configuration as required per documentation

* Fixing typo

"nesser" to "nessus"

* Provide a default instance name and a default scanner name

We're adding a sensible default name for the nessus scanner, and it 
makes sense to track those changes in the other resources used by the 
module.

* Make the elastic IP address optional, but use it by default

The EIP associated with the instance is called for by the documentation. 
I have a lot of conflicting evidence that says it's not required, so I'm 
making it optional but we will create the EIP by default.